Move MockingbirdAutomationCli into sep target
