+++ Mockingbird.xcodeproj/xcconfigs/MockingbirdCli.xcconfig
@@ -7,4 +7,4 @@ PRODUCT_NAME = MockingbirdCli
 TARGET_NAME = MockingbirdCli
 
 // Linking
-LD_RUNPATH_SEARCH_PATHS = @executable_path $(SRCROOT)/Sources/MockingbirdCli/Libraries $(TOOLCHAIN_DIR)/usr/lib/swift/macosx $(inherited)
+LD_RUNPATH_SEARCH_PATHS = @executable_path @executable_path/PackageFrameworks $(SRCROOT)/Sources/MockingbirdCli/Libraries $(TOOLCHAIN_DIR)/usr/lib/swift/macosx $(inherited)
