+++ Mockingbird.xcodeproj/xcconfigs/TestBase.xcconfig
@@ -6,7 +6,7 @@ CLANG_ENABLE_OBJC_WEAK = YES
 SWIFT_VERSION = 5.0
 
 // Compatibility
-MACOSX_DEPLOYMENT_TARGET = 10.14
+MACOSX_DEPLOYMENT_TARGET = 10.15
 
 // Linking
 LD_RUNPATH_SEARCH_PATHS = $(inherited) @loader_path/../Frameworks @loader_path/Frameworks
