+++ Sources/MockingbirdAutomation/Interop/Carthage.swift
@@ -15,8 +15,8 @@ public enum Carthage {
       "update",
       "--platform", platform.rawValue,
       "--use-xcframeworks",
-      "--verbose"
-    ], workingDirectory: project.parent()).runWithOutput()
+      "--verbose",
+    ], workingDirectory: project.parent()).run()
   }
   
   public static func build(platform: Platform = .all, project: Path) throws {
@@ -26,7 +26,7 @@ public enum Carthage {
       "--use-xcframeworks",
       "--no-skip-current",
       "--cache-builds",
-      "--verbose"
-    ], workingDirectory: project.parent()).runWithOutput()
+      "--verbose",
+    ], workingDirectory: project.parent()).run()
   }
 }
