+++ Sources/MockingbirdAutomation/Interop/CocoaPods.swift
@@ -3,6 +3,6 @@ import PathKit
 
 public enum CocoaPods {
   public static func install(workspace: Path) throws {
-    try Subprocess("pod", ["install"], workingDirectory: workspace.parent()).runWithOutput()
+    try Subprocess("pod", ["install"], workingDirectory: workspace.parent()).run()
   }
 }
