+++ Sources/MockingbirdAutomationCli/Commands/Test.swift
@@ -1,6 +1,7 @@
 import ArgumentParser
 import MockingbirdAutomation
 import PathKit
+import MockingbirdAutomation
 import Foundation
 
 struct Test: ParsableCommand {
