+++ Sources/MockingbirdAutomationCli/buildAndRun.sh
