+++ Sources/MockingbirdAutomationCli/main.swift
